test: refactor playground
